+++ Localization/frFR.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-01-07 19-00-12 ]] --
+-- [[ Exported at 2023-01-10 23-02-47 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --