+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Default values to the options
 - Tooltip data for Great Gourmand of The Ruby Feast
 - Option to hide the completion date of the achievements
+- Option to delay the initial Event Reminders on login (default 5 seconds)
 
 ### Changed
 - Optimized the Event Handling